Repo site initial upload.
This is the initial upload of the files the make up my site, in particular I'll be doing most edits to `art-gallery.html` related files and uploading here, then deploying to my site in a proper deployment fashion. It will still be manual, but ideally this repository will allow me to incorporate more automated deploys and testing.